requirements.txt
streamlit==1.37.1
json5==0.9.6
urllib3==2.2.2
websocket-client==1.8.0
pillow==10.4.0
openai==1.45.0
streamlit-extras==0.4.7
streamlit-drawable-canvas==0.9.0
opencv-python==4.10.0.84
torch==2.4.0
rembg==2.0.59